Add meeting notes from a meeting about strings
This was originally scheduled as @vmx wants to talk to @ribasushi as
he wants input from someone with a different perspective. We then inivted
@mvdan to the meeting.

No decisions were made (a subset of the team can't possibly make decisions
on such a topic).

One idea that came out of the meeting for further discussion is that we
could rename the `String` kind to `Unicode` to make its intention clearer.
Please see the meting minutes for more information.